streamlit/requirements.txt
numpy==1.23.2
pandas==1.4.3
Pillow==9.2.0
streamlit==1.12.0
tensorflow==2.9.1

